[refactor] ProviderとNotifierのフォルダを分離
