+++ project/bodquest_2023/lib/presentation/component/user/user_list.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 
-import '../../notifier/user/user_list_provider.dart';
+import '../../provider/user/user_list_provider.dart';
 import 'user_list_item.dart';
 
 class UserList extends ConsumerWidget {
