+++ project/bodquest_2023/lib/presentation/page/training/training_edit_page.dart
@@ -6,7 +6,7 @@ import '../../component/component_types.dart';
 import '../../component/control/number_textfield.dart';
 import '../../component/training/training_kind_dropdown.dart';
 import '../../notifier/datetime_notifier.dart';
-import '../../notifier/training/training_list_provider.dart';
+import '../../provider/training/training_list_provider.dart';
 import '../../notifier/text_notifier.dart';
 import '../../notifier/training/training_kind_notifier.dart';
 import '../../router/go_router.dart';
