+++ project/bodquest_2023/lib/presentation/page/weight/weight_page.dart
@@ -5,7 +5,7 @@ import '../../component/control/number_textfield.dart';
 import '../../component/weight/weight_list_button.dart';
 import '../../notifier/datetime_notifier.dart';
 import '../../notifier/text_notifier.dart';
-import '../../notifier/user/login_user_provider.dart';
+import '../../provider/user/login_user_provider.dart';
 import '../../notifier/weight/weight_list_provider.dart';
 import '../../router/go_router.dart';
 import '../../router/page_path.dart';
