+++ project/bodquest_2023/lib/presentation/provider/evaluation/evaluation_provider.dart
@@ -2,7 +2,7 @@ import 'package:flutter_riverpod/flutter_riverpod.dart';
 
 import '../../../domain/value/evaluation_rank.dart';
 import '../../state/evaluation/evaluation_state.dart';
-import 'evaluation_notifier.dart';
+import '../../notifier/evaluation/evaluation_notifier.dart';
 
 final evaluationNotifierProvider =
     StateNotifierProvider<EvaluationNotifier, EvaluationState>(
