+++ project/bodquest_2023/lib/presentation/provider/menu/menu_list_provider.dart
@@ -2,7 +2,7 @@ import 'package:flutter_riverpod/flutter_riverpod.dart';
 
 import '../../state/menu/menu_state.dart';
 import '../../usecaese_provider_module.dart';
-import 'menu_list_notifier.dart';
+import '../../notifier/menu/menu_list_notifier.dart';
 
 final menuListNotifierProvider =
     StateNotifierProvider<MenuListNotifier, AsyncValue<List<MenuState>>>(
