+++ project/bodquest_2023/lib/presentation/provider/user/login_user_provider.dart
@@ -2,7 +2,7 @@ import 'package:flutter_riverpod/flutter_riverpod.dart';
 
 import '../../state/user/loginuser_state.dart';
 import '../../usecaese_provider_module.dart';
-import 'login_user_notifier.dart';
+import '../../notifier/user/login_user_notifier.dart';
 
 final logInUserNotifierProvider =
     StateNotifierProvider<LogInUserNotifier, LogInUserState>(
